Move everything in src/ to scale/src/ to avoid name clashes in external projects
