+++ include/scale/scale.hpp
@@ -12,8 +12,8 @@
 #include <boost/throw_exception.hpp>
 #include <gsl/span>
 #include <outcome/outcome.hpp>
-#include "scale_decoder_stream.hpp"
-#include "scale_encoder_stream.hpp"
+#include "scale/scale_decoder_stream.hpp"
+#include "scale/scale_encoder_stream.hpp"
 
 namespace scale {
   /**
