+++ src/scale/common/buffer.hpp
