+++ src/scale/common/hexutil.cpp
@@ -3,7 +3,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "common/hexutil.hpp"
+#include "scale/common/hexutil.hpp"
+
 #include <boost/algorithm/hex.hpp>
 #include <boost/format.hpp>
 #include <gsl/span>
