+++ src/scale/common/hexutil.hpp
