+++ src/scale/common/visitor.hpp
