+++ src/scale/scale_encoder_stream.cpp
